+++ legacy_packages/unity-js-bridge/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @thirdweb-dev/unity-js-bridge
 
+## 0.6.38
+
+### Patch Changes
+
+- Updated dependencies [[`8d0c985`](https://github.com/thirdweb-dev/js/commit/8d0c985c45fc85b220784f841c1de7533ad8ebf8)]:
+  - @thirdweb-dev/chains@0.1.115
+  - @thirdweb-dev/sdk@4.0.89
+  - @thirdweb-dev/wallets@2.5.29
+  - @thirdweb-dev/auth@4.1.87
+
+## 0.6.37
+
+### Patch Changes
+
+- Updated dependencies [[`bd8fb7a`](https://github.com/thirdweb-dev/js/commit/bd8fb7a576ada21381625f1a241b10a6e68ef1f7)]:
+  - @thirdweb-dev/chains@0.1.114
+  - @thirdweb-dev/sdk@4.0.88
+  - @thirdweb-dev/wallets@2.5.28
+  - @thirdweb-dev/auth@4.1.86
+
+## 0.6.36
+
+### Patch Changes
+
+- Updated dependencies [[`4abd559`](https://github.com/thirdweb-dev/js/commit/4abd5591068bfa9217ec65e5830fba0cd14d11b9), [`642e7e9`](https://github.com/thirdweb-dev/js/commit/642e7e9f100dce2714e5eecd72cdc48aa9670004), [`32df45f`](https://github.com/thirdweb-dev/js/commit/32df45f9b07fa07e6f3ad218af095d721cacc76a)]:
+  - @thirdweb-dev/wallets@2.5.27
+  - @thirdweb-dev/chains@0.1.113
+  - @thirdweb-dev/sdk@4.0.87
+  - @thirdweb-dev/auth@4.1.85
+
+## 0.6.35
+
+### Patch Changes
+
+- [#3081](https://github.com/thirdweb-dev/js/pull/3081) [`c494af8`](https://github.com/thirdweb-dev/js/commit/c494af88d8a43d50bb95e1ab85148eb5a0bbf468) Thanks [@0xFirekeeper](https://github.com/0xFirekeeper)! - Unity 4.13.3
+
+- [#3062](https://github.com/thirdweb-dev/js/pull/3062) [`d874365`](https://github.com/thirdweb-dev/js/commit/d8743654e46f2687796f89194a407394a7ff36f6) Thanks [@0xFirekeeper](https://github.com/0xFirekeeper)! - Unity 4.13.2
+
+- Updated dependencies [[`2cd8a01`](https://github.com/thirdweb-dev/js/commit/2cd8a01ec07cc9dadf9570f807181be6b8d43047)]:
+  - @thirdweb-dev/chains@0.1.112
+  - @thirdweb-dev/sdk@4.0.86
+  - @thirdweb-dev/wallets@2.5.26
+  - @thirdweb-dev/auth@4.1.84
+
 ## 0.6.34
 
 ### Patch Changes